Adapt new TransExec changes in ASEIA

+++ src/aseia_car_sim/src/Bridge.cpp
@@ -19,7 +19,7 @@ const uint32_t UTM = 1;
 
 struct EventConfig : public BaseConfig {
   using TimeValueType = Value<double, 1>;
-  using PublisherIDValueType = Value<uint64_t, 1, 1, false>;
+  using PublisherIDValueType = Value<uint32_t, 1, 1, false>;
   using PositionValueType = Value<float, 2, 1, true>;
   using PositionScale = Scale<std::ratio<1>, UTM>;
 };
